requirements.txt
Django~=5.1
pillow~=10.4.0
pip~=23.2.1
asgiref~=3.8.1
sqlparse~=0.5.1
psycopg2~=2.9.9
djangorestframework~=3.15.2